feat(DTFS2-7298): improve facility mapping consistency (#3767)
