+++ trade-finance-manager-api/src/v1/rest-mappings/facility.js
@@ -1,22 +1,27 @@
 const mapFacilitySnapshot = require('./mappings/facilities/mapFacilitySnapshot');
 const mapFacilityTfm = require('./mappings/facilities/mapFacilityTfm');
-const mapGefFacility = require('./mappings/gef-facilities/mapGefFacility');
+const mapGefFacilitySnapshot = require('./mappings/gef-facilities/mapGefFacilitySnapshot');
 const isGefFacility = require('./helpers/isGefFacility');
 
+/**
+ * Maps a facility database object to a facility type throughout TFM-API and TFM-UI.
+ * The mappings include creating a facility snapshot that represents the current facility, with all amendments and tfm changes applied.
+ * @param {import('@ukef/dtfs2-common').TfmFacility} facility the full facility object from the database
+ * @param {import('@ukef/dtfs2-common').Deal} dealSnapshot the deal.dealSnapshot object from the database corresponding to the facility
+ * @param {import('@ukef/dtfs2-common').TfmDeal.tfm} dealTfm the deal.dealTfm object from the database corresponding to the facility
+ * @returns mapped facility object
+ */
 const facilityMapper = (facility, dealSnapshot, dealTfm) => {
   const facilityType = facility.facilitySnapshot.type;
+  const isGef = isGefFacility(facilityType);
 
-  if (isGefFacility(facilityType)) {
-    return mapGefFacility(facility, dealSnapshot, dealTfm);
-  }
+  const mappedFacilitySnapshot = isGef ? mapGefFacilitySnapshot(facility, dealSnapshot) : mapFacilitySnapshot(facility, dealSnapshot);
 
-  const result = {
+  return {
     _id: facility._id,
-    facilitySnapshot: mapFacilitySnapshot(facility, dealSnapshot.details),
+    facilitySnapshot: mappedFacilitySnapshot,
     tfm: mapFacilityTfm(facility.tfm, dealTfm, facility),
   };
-
-  return result;
 };
 
 module.exports = facilityMapper;
