+++ trade-finance-manager-api/src/v1/rest-mappings/mappings/gef-facilities/mapGefFacilities.js
@@ -1,13 +1,18 @@
-const mapGefFacility = require('./mapGefFacility');
+const facilityMapper = require('../../facility');
 
+/**
+ * Maps facilities from the database associated with a deal to their corresponding facility objects for use in TFM-UI and TFM-API.
+ * This returns facility objects that represents current facility states with all changes applied e.g. when amendments are added in TFM.
+ * @param {import('@ukef/dtfs2-common').Deal} dealSnapshot the deal.dealSnapshot object from the database corresponding to the facility
+ * @param {import('@ukef/dtfs2-common').TfmDeal.tfm} dealTfm the deal.tfm object from the database corresponding to the facility
+ * @returns mapped facility objects
+ */
 const mapGefFacilities = (dealSnapshot, dealTfm) => {
   const { facilities } = dealSnapshot;
 
   // Map facilities if only they exists
   if (facilities?.length) {
-    const mappedFacilities = facilities.map((facility) => mapGefFacility(facility, dealSnapshot, dealTfm));
-
-    return mappedFacilities;
+    return facilities.map((facility) => facilityMapper(facility, dealSnapshot, dealTfm));
   }
 
   return null;
